ffi: rework NativeEndian
Rework the NativeEndian trait to be fully self-reliant and provide all
implementations statically. The only requirement now is for
implementations to annotate whether they require a byte-swap or not.

This has the major disadvantage that all implementations are limited to
simple endian byte-swaps. However:

 - There is currently no target that Rust supports which does not fit
   this.
 - This is required to allow us to get access to NativeEndian methods in
   `const fn`.

Once Rust gains support for inherent `const fn` in traits, we can
finally call trait-methods as constructors for constants, and thus can
switch to a `NativeEndian` trait that allows for platforms outside of
little/big endian systems (if that is ever needed).

Signed-off-by: David Rheinsberg <david@readahead.eu>