osi/never: add the Never type
Add a new module to expose the unstable never-type `!` of Rust. Use a
small trick to ensure it can be used directly, rather than just as
return type of functions.

This hack exposes unstable details of Rust and as such should be used
with caution. However, this is highly useful for test-suites or manual
verification of Rust invariants. Hence, we provide it here.

Signed-off-by: David Rheinsberg <david@readahead.eu>