osi/meta: rework Field
Rework the Field trait again. Turn it into a marker trait and simplify
the helpers around it. Remove all helpers that can easily be misused and
drop the unused PinField for now.

Note that Stacked Borrows do not allow increasing the object size of a
reference. Instead, one must always ensure to retain raw pointers when
intending to reverse offset calculations. This is not an issue for Tree
Borrows, though. However, we prefer being compatible to Stacked Borrows,
so lets ensure `Field` is as well.

Signed-off-by: David Rheinsberg <david@readahead.eu>