osi/pin: introduce as_inner()
Provide an unsafe accessor for the wrapped pointer of pins. This is
documented to be accessible via transmute(), but an explicit helper goes
a long way.

This is particularly useful if the pin is part of a bigger structure,
and no mutable access is available. In those cases
Pin::into_inner_unchecked() cannot be used, as the source cannot be
moved. Instead, the new `as_inner()` will only borrow the original and
still provide access to the internals.

Signed-off-by: David Rheinsberg <david@readahead.eu>